Split DefinitionEntry into different files
